requirements.txt
packaging
torch
tqdm
absl-py
transformers
gradio
unstructured
pdf2image
pdfminer
pdfminer.six
opencv-python
unstructured_inference
layoutparser
onnxruntime
pikepdf
sentence-transformers
chromadb
flash-attn
json-repair
langchain-community
langchain
langchain_experimental
neo4j
pillow_heif
gradio
